+++ Sources/StreamChat/ChatClient.swift
@@ -631,7 +631,7 @@ public class ChatClient: @unchecked Sendable {
     public func loadAppSettings(
         completion: (@Sendable (Result<AppSettings, Error>) -> Void)? = nil
     ) {
-        apiClient.request(endpoint: .appSettings()) { [weak self] result in
+        apiClient.request(endpoint: .getApp()) { [weak self] result in
             switch result {
             case let .success(payload):
                 let appSettings = payload.asModel()