+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add user groups support [#4138](https://github.com/GetStream/stream-chat-swift/pull/4138)
 ### 🐞 Fixed
 - Fix sending typing events to channels without typing events capability [#4132](https://github.com/GetStream/stream-chat-swift/pull/4132)
+### 🔄 Changed
+- Use v2 app endpoint for app settings [#4133](https://github.com/GetStream/stream-chat-swift/pull/4133)
 
 ## StreamChatUI
 ### 🐞 Fixed